// frontend/src/client/types.gen.ts
export type ConversationSummary = {
    /**
     * Conversation Id
     */
    conversation_id: string;
    /**
     * Title
     */
    title: string | null;
    /**
     * Created At
     */
    created_at: string;
    /**
     * Message Count
     */
    message_count: number;
};